to_do function modification and file division.
